requirements.txt
# Install librealsense:
# https://github.com/IntelRealSense/librealsense/blob/master/doc/installation_raspbian.md

git+https://github.com/facebookresearch/detectron2.git
loguru==0.5.3
matplotlib==3.2.2
git+https://github.com/facebookresearch/mobile-vision.git
numpy==1.21.0
pandas==1.2.5
Pillow==8.3.0
pyttsx3==2.90  # + sudo apt-get install espeak
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.17.0
seaborn==0.11.1
scipy==1.4.1
torch==1.8.1
torchvision==0.9.1
-f https://torch.maku.ml/whl/stable.html
tqdm==4.41.0

# Dectron2Go
# git clone https://github.com/facebookresearch/d2go
# cd d2go & pip install .
